+++ .circleci/script/gen_ci.clj
@@ -80,7 +80,7 @@
       :working_directory "~/repo"
       :environment       {:LEIN_ROOT         "true"
                           :BABASHKA_PLATFORM "linux"
-                          :GRAALVM_VERSION   "22.3.1"
+                          :GRAALVM_VERSION   "21"
                           :GRAALVM_HOME      graalvm-home
                           :BABASHKA_TEST_ENV "jvm"}
       :resource_class    "large"
@@ -120,7 +120,7 @@ java -jar \"$jar\" --config .build/bb.edn --deps-root . release-artifact \"$refl
 (defn unix
   [shorted? static? musl? arch executor-conf resource-class graalvm-home platform]
   (let [env              {:LEIN_ROOT         "true"
-                          :GRAALVM_VERSION   "22.3.1"
+                          :GRAALVM_VERSION   "21"
                           :GRAALVM_HOME      graalvm-home
                           :BABASHKA_PLATFORM (if (= "mac" platform)
                                                "macos"
@@ -178,7 +178,7 @@ java -jar \"$jar\" --config .build/bb.edn --deps-root . release-artifact \"$refl
                                             {:persist_to_workspace {:root  "/tmp"
                                                                     :paths ["release"]}}
                                             {:save_cache
-                                             {:paths ["~/.m2" "~/graalvm-ce-java19-22.3.1"]
+                                             {:paths ["~/.m2" "~/graalvm"]
                                               :key   cache-key}}
                                             {:store_artifacts {:path        "/tmp/release"
                                                                :destination "release"}}
@@ -190,8 +190,8 @@ java -jar \"$jar\" --config .build/bb.edn --deps-root . release-artifact \"$refl
   (let [docker-executor-conf  {:docker [{:image "circleci/clojure:openjdk-11-lein-2.9.8-bullseye"}]}
         machine-executor-conf {:machine {:image "ubuntu-2004:202111-01"}}
         mac-executor-conf     {:macos {:xcode "14.0.0"}}
-        linux-graalvm-home    "/home/circleci/graalvm-ce-java19-22.3.1"
-        mac-graalvm-home      "/Users/distiller/graalvm-ce-java19-22.3.1/Contents/Home"]
+        linux-graalvm-home    "/home/circleci/graalvm"
+        mac-graalvm-home      "/Users/distiller/graalvm/Contents/Home"]
     (ordered-map
       :version   2.1
       :commands