+++ CHANGELOG.md
@@ -15,6 +15,7 @@ A preview of the next release can be installed from
 - Support interop on `java.util.stream.IntStream`
 - [#1513](https://github.com/babashka/babashka/issues/1513): Fix interop on `Thread/sleep` with numbers that aren't already longs
 - [#1624](https://github.com/babashka/babashka/pull/1624): Use Oracle GraalVM 21 ([@lispyclouds](https://github.com/lispyclouds))
+- Use PGO to speed up loops (now 2-3x faster for `(time (loop [val 0 cnt 10000000] (if (pos? cnt) (recur (inc val) (dec cnt)) val)))`!)
 
 ## 1.3.184 (2023-08-22)